// web/src/types.ts
export type JsonObject = Record<string, unknown>;
// ...
export interface Job {
  id: number;

  type_key: string;
  action_key: string;
  selector: JsonObject;
  params: JsonObject;
  status: number;
  last_run?: JobRun;
  created_at: string;
  updated_at: string;
}
// ...
export interface JobRun {
  id: number;
  job_id: number;
  job_type_key: string;
  job_action_key: string;
  account_id?: number;
  worker_id: string;
  attempt: number;
  result?: JsonObject;
  error_code?: string;
  error_message?: string;
  started_at: string;
  ended_at?: string;
}